Refactored, cleaned and commented code
It's still not very clean. *Sorry*. Some functions are in weird places,
but I don't know where to put them at this point.

Anyway... I think I'm done refactoring this game. The last commits have
changed nothing in functionality, and are for the sole purpose of making
the code more readable and *reusable*. Cheers.